Expose Pixar.Usd.PrimRange for stage traversal. * Mark Pixar.Usd.PrimRange with conformance to   CxxConvertibleToCollection. * Fix plugin paths plugInfo.json paths.

+++ Python/PyTrace/wrapCollector.cpp
@@ -43,7 +43,7 @@ using PythonKey = std::string;
 static double GetElapsedSeconds(TraceEvent::TimeStamp begin,
                                 TraceEvent::TimeStamp end) {
   if (begin > end) {
-    TF_CODING_ERROR("Invalid interval: begin=%zu, end=%zu", begin, end);
+    TF_CODING_ERROR("Invalid interval: begin=%zu, end=%zu", static_cast<size_t>(begin), static_cast<size_t>(end));
     return 0.0;
   }
   return ArchTicksToSeconds(end - begin);
